Adds multihinge loss and kplusone classifier where the main generator loss is either fm or opposite WGAN generator loss. Selects appropriate discriminator and generator based on image resolution and critic type arguments.
In transitioning from a multi-hinge loss on an auxiliary network to a kplusone network, a different loss is needed on the discriminator. For the AC case the correct class is enforced with Cramer-Singer (CS) on the K real classes for both real and generated samples. For the kplusone case, the correct class of the K+1 classes is enforce d for the real images, and the K+1th class (the fake class) is enforced for the generated images. For the generator, the the only difference between the AC and kplusone cases are the number of classes (K versus K+1), but either way the intended class is enforced. Then the 'main' losses of the AC case change: the D wasserstein hinge loss is not present, and for G either the loss is feature matching or the wasserstein loss with a sign flip (since the input is now the affinity for the fake class; previously the input was the affinity for real that came out of a binary discriminator).